// lib/types/request/SpecialRequestInterface.ts
export interface SpecialRequestInterface {
    name?: string
    serviceType?: string
    frequency?: string
    rate?: string
    lan?: string
    log?: string
    address?: string
    image?: any
    condition?: string
    type?: string
    categoryId?: string
    subcategoryId?: string
    description?: string
}